feat: add module structure in project
