+++ js/constants.js
@@ -1,0 +1,43 @@
+const MAX_AVATAR_COUNT = 6;
+const MAX_PHOTO_COUNT = 25;
+const MAX_COMMENT_COUNT = 999;
+const LIKES_RANGE = [15, 200];
+const COMMENTS_RANGE = [0, 30];
+
+const PHOTO_DESCRIPTIONS = [
+  'Вдохновение и красота вокруг нас',
+  'Мгновение, остановленное во времени',
+  'Искусство видеть детали',
+  'Жизнь в её лучших проявлениях',
+  'Простота и гармония',
+  'Магия момента'
+];
+
+const COMMENT_NAMES = [
+  'Вайолет',
+  'Гилберт',
+  'Клаудия',
+  'Каттлея',
+  'Эрика',
+  'Айрис'
+];
+
+const COMMENT_MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
+
+export {
+  MAX_AVATAR_COUNT,
+  MAX_PHOTO_COUNT,
+  MAX_COMMENT_COUNT,
+  LIKES_RANGE,
+  COMMENTS_RANGE,
+  PHOTO_DESCRIPTIONS,
+  COMMENT_NAMES,
+  COMMENT_MESSAGES
+};
